Do not destroy object variables if _SSL_opening is true
In situations where the first start_SSL call (server side) fails due to
SSL_WANT_READ, we're supposed to be able to re-try. Clearing all object
variables on a retry can be disastrous as it triggers unwanted DESTROY events.

Added unit test to show this. It will sometimes succeed though 10 iterations
seems to be enough to trigger it most times.