Fix Net::SSLeay version string comparisons
The value of $Net::SSLeay::VERSION is subjected to several numerical
comparisons, all of which use the standard relational operators. These
tests do not take into account the innate flexibility of Perl version
strings and will trigger warnings in the numeric category for values
such as "1.86_11".

Fix this by using the version module to parse Net::SSLeay's version
string. The resulting object will overload the operators so as to do the
right thing.